requirements.txt
click==7.1.2
h11==0.8
mysql==0.0.2
mysqlclient==2.0.1
uvicorn==0.13.1
aiofiles==0.5.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
httpcore==0.10.2
httptools==0.1.1
httpx==0.14.3
idna==2.10
loguru==0.5.1
mailchimp3==3.0.14
numpy==1.19.1
pandas==1.1.1
psycopg2==2.8.6
pycountry==20.7.3
pydantic==1.6.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
starlette==0.13.6
urllib3==1.25.10
websockets==8.1
texthero==1.0.9
pyodbc==4.0.30
fbprophet==0.7.1
sklearn==0.0
dask==2.26.0
prefect==0.14.6